Projeto integrado com Banco de Dados PostgreSQL utilizando o padrão MVC e DTO

+++ src/main/java/com/masterbare/projetobackendprimeiroexemplo/services/ProdutoService.java
@@ -2,10 +2,16 @@ package com.masterbare.projetobackendprimeiroexemplo.services;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
+
+import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import com.masterbare.projetobackendprimeiroexemplo.model.Produto;
+import com.masterbare.projetobackendprimeiroexemplo.model.exception.ResourceNotFoundException;
 import com.masterbare.projetobackendprimeiroexemplo.repository.ProdutoRepository;
+import com.masterbare.projetobackendprimeiroexemplo.shared.ProdutoDTO;
+//import com.masterbare.projetobackendprimeiroexemplo.view.controller.*;
 
 @Service
 public class ProdutoService {
@@ -17,10 +23,16 @@ public class ProdutoService {
      * Método para retornar uma lista de produtos
      * @return Lista dos produtos
      */
-    public List<Produto> obterTodos() {
+    public List<ProdutoDTO> obterTodos() {
         // Colocar regra aqui caso tenha ...
-        //return produtoRepository.obterTodos();
-        return produtoRepository.findAll();
+        // eturn produtoRepository.obterTodos();
+        // Retorna uma lista de produto model.
+        List<Produto> produtos = produtoRepository.findAll();
+
+        return produtos.stream()
+        .map(produto -> new ModelMapper().map(produto, ProdutoDTO.class))
+        .collect(Collectors.toList());
+
     }
 
     /**
@@ -28,9 +40,21 @@ public class ProdutoService {
      * @param id do produto que será localizado. 
      * @return Retorna um produto caso seja encontrado.
      */
-    public Optional<Produto> obterPorId(Long id) {
+    public Optional<ProdutoDTO> obterPorId(Long id) {
         //return produtoRepository.obterPorId(id);
-        return produtoRepository.findById(id);
+        // Obtendo optional de produto pelo id.
+        Optional<Produto> produto =  produtoRepository.findById(id);
+
+        // Se não encontrar, lança exception.
+        if(produto.isEmpty()){
+            throw new ResourceNotFoundException("Produto com id: " + id + " não encontrado.");
+        }
+
+        // Convertendo o meu optional de produto em um produtoDTO
+        ProdutoDTO dto = new ModelMapper().map(produto.get(), ProdutoDTO.class);
+
+        // Criando e retornando um optional de produtoDTO.
+        return Optional.of(dto);
     }
 
      /**
@@ -38,10 +62,23 @@ public class ProdutoService {
      * @param produto que será adicionado.
      * @return Retorna o produto que foi adicionado na lista.
      */
-    public Produto adicionar(Produto produto) {
-        // Poderia ter alguma regra de negócio para validar o produto.
-        //return produtoRepository.adicionar(produto);
-        return produtoRepository.save(produto);
+    public ProdutoDTO adicionar(ProdutoDTO produtoDto) {
+        // Removendo o id para conseguir fazer o cadastro.
+        produtoDto.setId(null);
+
+        // Criar um objeto de mapeamento.
+        ModelMapper mapper = new ModelMapper();
+
+        // Converter o nosso ProdutoDTO em um Produto.
+        Produto produto = mapper.map(produtoDto, Produto.class);
+
+        // Salvar o Produto no banco.
+        produto = produtoRepository.saveAndFlush(produto);
+
+        produtoDto.setId(produto.getId());
+
+        // Retornar o ProdutoDTO atualizado.
+        return produtoDto;
     }
 
     /**
@@ -49,8 +86,15 @@ public class ProdutoService {
      * @param id do produto a ser deletado.
      */
     public void deletar(Long id) {
-        // Aqui poderia ter alguma lógica de validação.
-        //produtoRepository.deletar(id);
+        // Verificar se o produto existe
+        Optional<Produto> produto = produtoRepository.findById(id);
+
+        // Se não existir lança uma exception
+        if(produto.isEmpty()){
+            throw new ResourceNotFoundException("Não foi possível deletar o produto com o id: " + id + " - Produto não existe.");
+        }
+
+        // Deleta o produto pelo id.
         produtoRepository.deleteById(id);
     }
 
@@ -60,13 +104,23 @@ public class ProdutoService {
      * @param id do produto.
      * @return Retorna o produto após atualizar a lista.
      */
-    public Produto atualizar(Long id, Produto produto) {
+    public ProdutoDTO atualizar(Long id, ProdutoDTO produtoDto) {
        
-        // Ter alguma validação no ID.
-        produto.setId(id);
+        // Passar o id para o produto Dto.
+        produtoDto.setId(id);
 
-        //return produtoRepository.atualizar(produto);
-        return produtoRepository.saveAndFlush(produto);
+        // Criar um objeto de mapeamento.
+        ModelMapper mapper = new ModelMapper();
+
+        // Converter o ProdutoDTO em um produto.
+        Produto produto = mapper.map(produtoDto, Produto.class);
+
+        // Atualizar o produto no Banco de dados.
+        produtoRepository.saveAndFlush(produto);
+        
+        // Retornar o produtoDto atualizado.
+        return produtoDto;
+
     }
 
 }
